Move around classes to better packages
