+++ library/item/extension/build.gradle
@@ -4,8 +4,8 @@ plugins {
 
 qslModule {
 	name = "Quilt Item Extension API"
-	moduleName = "extension"
-	id = "quilt_extension"
+	moduleName = "item_extension"
+	id = "quilt_item_extension"
 	description = "An API for extending various vanilla items that would be difficult to otherwise. " +
 			"Also provides relevant events for these items."
 	library = "item"
@@ -15,9 +15,9 @@ qslModule {
 		}
 	}
 	injectedInterface("net/minecraft/class_1753") {
-		values = ["org/quiltmc/qsl/item/extension/impl/BowExtensions"]
+		values = ["org/quiltmc/qsl/item/extension/api/bow/BowExtensions"]
 	}
 	injectedInterface("net/minecraft/class_1764") {
-		values = ["org/quiltmc/qsl/item/extension/impl/CrossbowExtensions"]
+		values = ["org/quiltmc/qsl/item/extension/api/crossbow/CrossbowExtensions"]
 	}
 }
