+++ library/item/extension/src/main/java/org/quiltmc/qsl/item/extension/api/crossbow/CrossbowExtensions.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.quiltmc.qsl.item.extension.impl;
+package org.quiltmc.qsl.item.extension.api.crossbow;
 
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.item.CrossbowItem;
@@ -22,16 +22,14 @@ import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import org.jetbrains.annotations.ApiStatus;
 import org.quiltmc.qsl.base.api.util.InjectedInterface;
-import org.quiltmc.qsl.item.extension.api.ProjectileModifyingCrossbowItem;
-import org.quiltmc.qsl.item.extension.api.ShotProjectileEvents;
-import org.quiltmc.qsl.item.extension.mixin.CrossbowItemAccessor;
+import org.quiltmc.qsl.item.extension.mixin.crossbow.CrossbowItemAccessor;
 
 /**
  * An interface to implement for all custom crossbows in Quilt.
  * <p>
  * This is meant to be used on a {@link net.minecraft.item.CrossbowItem} class.</p>
  * <p>
- * In order to modify the projectile shot from your crossbow, implementing and registering {@link ShotProjectileEvents.ModifyProjectileFromCrossbow} is recommended.</p>
+ * In order to modify the projectile shot from your crossbow, implementing and registering {@link CrossbowShotProjectileEvents.ModifyProjectileFromCrossbow} is recommended.</p>
  *
  * @see ProjectileModifyingCrossbowItem
  */
