+++ library/item/extension/src/main/java/org/quiltmc/qsl/item/extension/api/crossbow/ProjectileModifyingCrossbowItem.java
@@ -14,24 +14,24 @@
  * limitations under the License.
  */
 
-package org.quiltmc.qsl.item.extension.api;
+package org.quiltmc.qsl.item.extension.api.crossbow;
 
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.projectile.PersistentProjectileEntity;
 import net.minecraft.item.CrossbowItem;
 import net.minecraft.item.ItemStack;
 import org.jetbrains.annotations.NotNull;
-import org.quiltmc.qsl.item.extension.impl.CrossbowExtensions;
+import org.quiltmc.qsl.item.extension.api.bow.BowShotProjectileEvents;
 
 /**
  * This is the default implementation for {@link CrossbowExtensions}, allowing for the easy creation of new crossbows with no new modded functionality.
  * <p>
  * This crossbow automatically registers itself to modify its projectiles with {@link ProjectileModifyingCrossbowItem#onProjectileShot(ItemStack, ItemStack, LivingEntity, PersistentProjectileEntity)}</p>
  */
-public class ProjectileModifyingCrossbowItem extends CrossbowItem implements ShotProjectileEvents.ModifyProjectileFromCrossbow {
+public class ProjectileModifyingCrossbowItem extends CrossbowItem implements CrossbowShotProjectileEvents.ModifyProjectileFromCrossbow {
 	public ProjectileModifyingCrossbowItem(Settings settings) {
 		super(settings);
-		ShotProjectileEvents.CROSSBOW_MODIFY_SHOT_PROJECTILE.register(this);
+		CrossbowShotProjectileEvents.CROSSBOW_MODIFY_SHOT_PROJECTILE.register(this);
 	}
 
 	@Override
