+++ library/item/extension/src/main/java/org/quiltmc/qsl/item/extension/mixin/ProjectileUtilMixin.java
@@ -21,8 +21,8 @@ import net.minecraft.entity.projectile.ProjectileUtil;
 import net.minecraft.item.Item;
 import net.minecraft.item.Items;
 import net.minecraft.util.Hand;
-import org.quiltmc.qsl.item.extension.impl.BowExtensions;
-import org.quiltmc.qsl.item.extension.impl.CrossbowExtensions;
+import org.quiltmc.qsl.item.extension.api.bow.BowExtensions;
+import org.quiltmc.qsl.item.extension.api.crossbow.CrossbowExtensions;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
