+++ library/item/extension/src/main/java/org/quiltmc/qsl/item/extension/mixin/bow/EntityUseBowMixin.java
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package org.quiltmc.qsl.item.extension.mixin;
+package org.quiltmc.qsl.item.extension.mixin.bow;
 
 import net.minecraft.entity.mob.AbstractSkeletonEntity;
 import net.minecraft.item.RangedWeaponItem;
-import org.quiltmc.qsl.item.extension.impl.BowExtensions;
+import org.quiltmc.qsl.item.extension.api.bow.BowExtensions;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
