+++ library/item/extension/src/main/java/org/quiltmc/qsl/item/extension/mixin/bow/client/SkeletonEntityModelMixin.java
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package org.quiltmc.qsl.item.extension.mixin.client;
+package org.quiltmc.qsl.item.extension.mixin.bow.client;
 
 import net.minecraft.client.render.entity.model.SkeletonEntityModel;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
-import org.quiltmc.qsl.item.extension.impl.BowExtensions;
+import org.quiltmc.qsl.item.extension.api.bow.BowExtensions;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Redirect;
@@ -27,13 +27,13 @@ import org.spongepowered.asm.mixin.injection.Redirect;
 @Mixin(SkeletonEntityModel.class)
 public abstract class SkeletonEntityModelMixin {
 	// Allows Skeletons to visually shoot custom bows by returning true
-	@Redirect(method = "animateModel", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;isOf(Lnet/minecraft/item/Item;)Z"))
+	@Redirect(method = "animateModel(Lnet/minecraft/entity/mob/MobEntity;FFF)V", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;isOf(Lnet/minecraft/item/Item;)Z"))
 	private boolean animateModel(ItemStack heldItemStack, Item item) {
 		return heldItemStack.getItem() instanceof BowExtensions;
 	}
 
 	// Allows Skeletons to visually shoot custom bows by returning true
-	@Redirect(method = "setAngles", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;isOf(Lnet/minecraft/item/Item;)Z"))
+	@Redirect(method = "setAngles(Lnet/minecraft/entity/mob/MobEntity;FFFFF)V", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;isOf(Lnet/minecraft/item/Item;)Z"))
 	private boolean setAngles(ItemStack heldItemStack, Item item) {
 		return heldItemStack.getItem() instanceof BowExtensions;
 	}
