+++ library/item/extension/src/main/java/org/quiltmc/qsl/item/extension/mixin/client/HeldItemRendererMixin.java
@@ -20,8 +20,8 @@ import net.minecraft.client.render.item.HeldItemRenderer;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
-import org.quiltmc.qsl.item.extension.impl.BowExtensions;
-import org.quiltmc.qsl.item.extension.impl.CrossbowExtensions;
+import org.quiltmc.qsl.item.extension.api.bow.BowExtensions;
+import org.quiltmc.qsl.item.extension.api.crossbow.CrossbowExtensions;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Redirect;
