+++ library/item/extension/src/main/resources/assets/quilt_item_extension/icon.png
