--- library/item/extension/src/main/resources/quilt_extension.mixins.json
@@ -1,24 +1,0 @@
-{
-  "required": true,
-  "package": "org.quiltmc.qsl.item.extension.mixin",
-  "compatibilityLevel": "JAVA_17",
-  "mixins": [
-    "AbstractSkeletonEntityMixin",
-    "BowAttackGoalMixin",
-    "BowAttackMixin",
-    "BowItemMixin",
-    "CrossbowItemAccessor",
-    "CrossbowItemMixin",
-    "EntityUseBowMixin",
-    "EntityUseCrossbowMixin",
-    "ProjectileUtilMixin"
-  ],
-  "client": [
-    "client.HeldItemRendererMixin",
-    "client.PlayerEntityRendererMixin",
-    "client.SkeletonEntityModelMixin"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
